Fix: no hidding button and filter by int

+++ app/src/main/java/com/iberdrola/practicas2026/davidsc/domain/model/InvoiceFilter.kt
@@ -11,7 +11,7 @@ import java.time.LocalDate
 data class InvoiceFilter(
     val desde: LocalDate? = null,
     val hasta: LocalDate? = null,
-    val importeMin: Double? = null,
-    val importeMax: Double? = null,
+    val importeMin: Int? = null,
+    val importeMax: Int? = null,
     val estados: Set<String> = emptySet()
 )
